requirements.txt
cryptography>=41.0.0
argon2-cffi>=21.3.0
rich>=13.4.0
